+++ src/client-utils.js
@@ -4,12 +4,12 @@
  * @class layer.ClientUtils
  */
 
-const LayerParser = require('layer-patch');
 const uuid = require('uuid');
-const atob = typeof window === 'undefined' ? require('atob') : window.atob;
 
-/* istanbul ignore next */
-const LocalFileReader = typeof window === 'undefined' ? require('filereader') : window.FileReader;
+exports.atob = typeof atob === 'undefined' ? global.getNativeSupport('atob') : atob.bind(window);
+exports.btoa = typeof btoa === 'undefined' ? global.getNativeSupport('btoa') : btoa.bind(window);
+const LocalFileReader = typeof FileReader === 'undefined' ? global.getNativeSupport('FileReader') : FileReader;
+
 
 /**
  * Generate a random UUID
@@ -18,7 +18,6 @@ const LocalFileReader = typeof window === 'undefined' ? require('filereader') : 
  * @return {string}
  */
 exports.generateUUID = uuid.v4;
-
 
 /**
  * Returns the 'type' portion of a Layer ID.
@@ -97,88 +96,6 @@ exports.sortBy = (inArray, fn, reverse) => {
  */
 exports.clone = obj => JSON.parse(JSON.stringify(obj));
 
-/**
- * Execute this function immediately after current processing is complete.
- *
- * A depth of up to 10 is allowed.  That means that functions you schedule using defer
- * can in turn schedule further actions.  The original actions are depth = 0; the actions scheduled
- * by your actions are depth = 1.  These new actions may in turn schedule further actions, which happen at depth = 3.
- * But to avoid infinite loops, if depth reaches 10, it clears the queue and ignores them.
- *
- * @method defer
- * @param {Function} f
- */
-let setImmediateId = 0,
-  setImmediateDepth = 0,
-
-  // Have we scheduled the queue to be processed? If not, this is false
-  setImmediateIsPending = false,
-
-  // Queue of functions to call and depth integers
-  setImmediateQueue = [];
-
-// If a setImmediate callback itself calls setImmediate which in turn calls setImmediate, at what point do we suspect we have an infinite loop?
-// A depth of 10 is currently considered OK, but this may need to be increased.
-const setImmediateMaxDepth = 10;
-
-// Process all callbacks in the setImmediateQueue
-function setImmediateProcessor() {
-  // Processing the queue is no longer scheduled; clear any scheduling info.
-  setImmediateIsPending = false;
-  clearTimeout(setImmediateId);
-  setImmediateId = 0;
-
-  // Our initial depth is depth 0
-  setImmediateDepth = 0;
-  setImmediateQueue.push(setImmediateDepth);
-
-  // Process all functions and depths in the queue starting always with the item at index 0,
-  // and removing them from the queue before processing them.
-  while (setImmediateQueue.length) {
-    const item = setImmediateQueue.shift();
-    if (typeof item === 'function') {
-      try {
-        item();
-      } catch (err) {
-        console.error(err);
-      }
-    } else if (item >= setImmediateMaxDepth) {
-      setImmediateQueue = [];
-      console.error('Layer Error: setImmediate Max Queue Depth Exceded');
-    }
-  }
-}
-
-// Schedule the function to be called by adding it to the queue, and setting up scheduling if its needed.
-function defer(func) {
-  if (typeof func !== 'function') throw new Error('Function expected in defer');
-  setImmediateQueue.push(func);
-
-  // If postMessage has not already been called, call it
-  if (!setImmediateIsPending) {
-    setImmediateIsPending = true;
-    if (typeof document !== 'undefined') {
-      window.postMessage({ type: 'layer-set-immediate' }, '*');
-    } else {
-      // React Native reportedly lacks a document, and throws errors on the second parameter
-      window.postMessage({ type: 'layer-set-immediate' });
-    }
-
-    // Having seen scenarios where postMessage failed to trigger, set a backup using setTimeout that will be canceled
-    // if postMessage is succesfully called.
-    setImmediateId = setTimeout(setImmediateProcessor, 0);
-  }
-}
-
-// For Unit Testing
-defer.flush = () => setImmediateProcessor();
-
-addEventListener('message', (event) => {
-  if (event.data.type !== 'layer-set-immediate') return;
-  setImmediateProcessor();
-});
-
-exports.defer = defer;
 
 /**
  * URL Decode a URL Encoded base64 string
@@ -205,7 +122,7 @@ exports.decode = (str) => {
     default:
       throw new Error('Illegal base64url string!');
   }
-  return atob(output);
+  return exports.atob(output);
 };
 
 /**
@@ -289,7 +206,7 @@ exports.blobToBase64 = (blob, callback) => {
 exports.base64ToBlob = (b64Data, contentType) => {
   try {
     const sliceSize = 512;
-    const byteCharacters = atob(b64Data);
+    const byteCharacters = exports.atob(b64Data);
     const byteArrays = [];
     let offset;
 
@@ -323,7 +240,7 @@ exports.base64ToBlob = (b64Data, contentType) => {
  * @param {String} str
  * @return {String}
  */
-exports.utoa = str => btoa(unescape(encodeURIComponent(str)));
+exports.utoa = str => exports.btoa(unescape(encodeURIComponent(str)));
 
 /**
  * Does window.atob() in a way that can decode data from utoa()
@@ -334,7 +251,7 @@ exports.utoa = str => btoa(unescape(encodeURIComponent(str)));
  * @param {String} str
  * @return {String}
  */
-exports.atou = str => decodeURIComponent(escape(atob(str)));
+exports.atou = str => decodeURIComponent(escape(exports.atob(str)));
 
 
 /**
@@ -357,54 +274,18 @@ exports.fetchTextFromFile = (file, callback) => {
 };
 
 
-let parser;
-
 /**
- * Creates a LayerParser
+ * Execute this function immediately after current processing is complete (setImmediate replacement).
  *
- * @method
- * @private
- * @param {Object} request - see layer.ClientUtils.layerParse
+ * A depth of up to 10 is allowed.  That means that functions you schedule using defer
+ * can in turn schedule further actions.  The original actions are depth = 0; the actions scheduled
+ * by your actions are depth = 1.  These new actions may in turn schedule further actions, which happen at depth = 3.
+ * But to avoid infinite loops, if depth reaches 10, it clears the queue and ignores them.
+ *
+ * @method defer
+ * @param {Function} f
  */
-function createParser(request) {
-  request.client.once('destroy', () => (parser = null));
-
-  parser = new LayerParser({
-    camelCase: true,
-    getObjectCallback: id => request.client.getObject(id),
-    createObjectCallback: (id, obj) => request.client._createObject(obj),
-    propertyNameMap: {
-      Conversation: {
-        unreadMessageCount: 'unreadCount',
-      },
-      Identity: {
-        presence: '_presence',
-      },
-    },
-    changeCallbacks: {
-      Message: {
-        all: (updateObject, newValue, oldValue, paths) => {
-          updateObject._handlePatchEvent(newValue, oldValue, paths);
-        },
-      },
-      Conversation: {
-        all: (updateObject, newValue, oldValue, paths) => {
-          updateObject._handlePatchEvent(newValue, oldValue, paths);
-        },
-      },
-      Channel: {
-        all: (updateObject, newValue, oldValue, paths) => {
-          updateObject._handlePatchEvent(newValue, oldValue, paths);
-        },
-      },
-      Identity: {
-        all: (updateObject, newValue, oldValue, paths) => {
-          updateObject._handlePatchEvent(newValue, oldValue, paths);
-        },
-      },
-    },
-  });
-}
+exports.defer = require('./utils/defer');
 
 /**
  * Run the Layer Parser on the request.
@@ -421,16 +302,15 @@ function createParser(request) {
  *      });
  *
  * @method
+ * @deprecated Use 'utils/layer-parser' instead
  * @param {Object} request - layer-patch parameters
  * @param {Object} request.object - Object being updated  by the operations
  * @param {string} request.type - Type of object being updated
  * @param {Object[]} request.operations - Array of change operations to perform upon the object
  * @param {layer.Client} request.client
  */
-exports.layerParse = (request) => {
-  if (!parser) createParser(request);
-  parser.parse(request);
-};
+exports.layerParse = require('./utils/layer-parser');
+
 
 /**
  * Object comparison.