+++ src/client.js
@@ -614,7 +614,7 @@ Client.prototype._scheduleCheckAndPurgeCacheAt = 0;
  * @static
  * @type {String}
  */
-Client.version = '3.2.2';
+Client.version = '3.2.3';
 
 /**
  * Any Conversation or Message that is part of a Query's results are kept in memory for as long as it